Some experimental framework that makes one gadget per team.
(Gadgets get discarded silently when not needed.)

Won't keep this cause it multiplies the number of callins by
the amount of AI controlled teams, and the teams still have
to filter the events. (Compare to team[unitTeam].UnitCreated(...))

Also it's hard to share common data (e.g. some analyzation of
map/flags) between teams this way.