Don't give the first order as non-queued order.
It makes most sense to do give it as queued order because in most
other code I assume a unit will be at his destination before he will
start to go to a new destination.

In other words, without this, units would often "skip" some waypoints
in a second pathing, if they had to fight to reach it, for example.